Move setUserType and setUserSignedIn to currentUserRedux instead of progressRedux
