+++ apps/src/code-studio/components/progress/ScriptOverviewHeader.jsx
@@ -5,7 +5,7 @@ import {connect} from 'react-redux';
 import ProtectedStatefulDiv from '@cdo/apps/templates/ProtectedStatefulDiv';
 import PlcHeader from '@cdo/apps/code-studio/plc/header';
 import {ViewType} from '@cdo/apps/code-studio/viewAsRedux';
-import {SignInState} from '@cdo/apps/code-studio/progressRedux';
+import {SignInState} from '@cdo/apps/templates/currentUserRedux';
 import ScriptAnnouncements from './ScriptAnnouncements';
 import {
   announcementShape,
