+++ apps/src/code-studio/initApp/loadApp.js
@@ -4,7 +4,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {TestResults} from '@cdo/apps/constants';
 import {getStore} from '../redux';
-import {SignInState, mergeProgress} from '../progressRedux';
+import {mergeProgress} from '../progressRedux';
+import {SignInState} from '@cdo/apps/templates/currentUserRedux';
 import {setVerified} from '@cdo/apps/code-studio/verifiedTeacherRedux';
 import {files} from '@cdo/apps/clientApi';
 var renderAbusive = require('./renderAbusive');
