+++ apps/src/code-studio/progressRedux.js
@@ -3,7 +3,6 @@
  */
 import $ from 'jquery';
 import _ from 'lodash';
-import {makeEnum} from '../utils';
 import {mergeActivityResult, activityCssClass} from './activityUtils';
 import {LevelStatus, LevelKind} from '@cdo/apps/util/sharedConstants';
 import {TestResults} from '@cdo/apps/constants';
@@ -20,8 +19,6 @@ const MERGE_PEER_REVIEW_PROGRESS = 'progress/MERGE_PEER_REVIEW_PROGRESS';
 const UPDATE_FOCUS_AREAS = 'progress/UPDATE_FOCUS_AREAS';
 const SHOW_TEACHER_INFO = 'progress/SHOW_TEACHER_INFO';
 const DISABLE_POST_MILESTONE = 'progress/DISABLE_POST_MILESTONE';
-const SET_USER_SIGNED_IN = 'progress/SET_USER_SIGNED_IN';
-const SET_USER_TYPE = 'progress/SET_USER_TYPE';
 const SET_IS_HOC_SCRIPT = 'progress/SET_IS_HOC_SCRIPT';
 const SET_IS_AGE_13_REQUIRED = 'progress/SET_IS_AGE_13_REQUIRED';
 const SET_IS_SUMMARY_VIEW = 'progress/SET_IS_SUMMARY_VIEW';
@@ -30,8 +27,6 @@ const SET_STUDENT_DEFAULTS_SUMMARY_VIEW =
 const SET_CURRENT_STAGE_ID = 'progress/SET_CURRENT_STAGE_ID';
 const SET_SCRIPT_COMPLETED = 'progress/SET_SCRIPT_COMPLETED';
 const SET_STAGE_EXTRAS_ENABLED = 'progress/SET_STAGE_EXTRAS_ENABLED';
-
-export const SignInState = makeEnum('Unknown', 'SignedIn', 'SignedOut');
 
 const PEER_REVIEW_ID = -1;
 
@@ -55,8 +50,6 @@ const initialState = {
   peerReviewStage: null,
   peerReviewsPerformed: [],
   showTeacherInfo: false,
-  signInState: SignInState.Unknown,
-  userType: 'unknown',
   postMilestoneDisabled: false,
   isHocScript: null,
   isAge13Required: false,
@@ -153,22 +146,6 @@ export default function reducer(state = initialState, action) {
     return {
       ...state,
       postMilestoneDisabled: true
-    };
-  }
-
-  if (action.type === SET_USER_SIGNED_IN) {
-    return {
-      ...state,
-      signInState: action.isSignedIn
-        ? SignInState.SignedIn
-        : SignInState.SignedOut
-    };
-  }
-
-  if (action.type === SET_USER_TYPE) {
-    return {
-      ...state,
-      userType: action.userType
     };
   }
 
@@ -431,11 +408,6 @@ export const updateFocusArea = (changeFocusAreaPath, focusAreaStageIds) => ({
 export const showTeacherInfo = () => ({type: SHOW_TEACHER_INFO});
 
 export const disablePostMilestone = () => ({type: DISABLE_POST_MILESTONE});
-export const setUserSignedIn = isSignedIn => ({
-  type: SET_USER_SIGNED_IN,
-  isSignedIn
-});
-export const setUserType = userType => ({type: SET_USER_TYPE, userType});
 export const setIsHocScript = isHocScript => ({
   type: SET_IS_HOC_SCRIPT,
   isHocScript
