+++ apps/src/craft/designer/craft.js
@@ -25,7 +25,7 @@ import Sounds from '../../Sounds';
 import {TestResults} from '../../constants';
 import trackEvent from '../../util/trackEvent';
 import {captureThumbnailFromCanvas} from '../../util/thumbnail';
-import {SignInState} from '../../code-studio/progressRedux';
+import {SignInState} from '@cdo/apps/templates/currentUserRedux';
 
 const MEDIA_URL = '/blockly/media/craft/';
