+++ apps/src/templates/SignInOrAgeDialog.js
@@ -6,7 +6,7 @@ import BaseDialog from '@cdo/apps/templates/BaseDialog';
 import color from '@cdo/apps/util/color';
 import Button from '@cdo/apps/templates/Button';
 import AgeDropdown from '@cdo/apps/templates/AgeDropdown';
-import {SignInState} from '@cdo/apps/code-studio/progressRedux';
+import {SignInState} from '@cdo/apps/templates/currentUserRedux';
 import i18n from '@cdo/locale';
 import {reload} from '@cdo/apps/utils';
 import {environmentSpecificCookieName} from '@cdo/apps/code-studio/utils';
