+++ apps/test/unit/code-studio/components/ShareDialogTest.js
@@ -2,7 +2,7 @@ import {assert} from '../../../util/reconfiguredChai';
 import React from 'react';
 import {shallow} from 'enzyme';
 import {UnconnectedShareDialog as ShareDialog} from '@cdo/apps/code-studio/components/ShareDialog';
-import {SignInState} from '@cdo/apps/code-studio/progressRedux';
+import {SignInState} from '@cdo/apps/templates/currentUserRedux';
 
 describe('ShareDialog', () => {
   it('renders our signed in version when signed in', () => {
